requirements.txt
solo-cli
anthropic
opencv-python-headless
elevenlabs
python-dotenv
rerun-sdk
pexpect
